DRF-Projects/06_MyBoard/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
whitenoise==6.8.2
